requirements.txt
altair==5.1.2
duckdb==0.9.1
networkx==3.1
pandas==1.5.3
pandas_profiling==3.2.0
protobuf==4.25.0
pyvis==0.3.2
Requests
simplejson==3.19.2
streamlit==1.27.2
streamlit_pandas_profiling==0.1.3
google-cloud-bigquery
db-dtypes
tqdm
stqdm

# panel==0.10.0
# holowviews==1.14.6
# bokeh==2.4.0

streamlit-observable
inequality
pygini

kaleido
web3

pyvis
python-dateutil